backend/requirements.txt
# Core
Django==5.0.0                     # Main web framework
djangorestframework==3.14.0       # For building REST APIs
django-cors-headers==4.3.1        # Handle CORS headers
django-rest-knox==4.2.0          # Better token authentication than JWT

# Channels for WebSocket
daphne==4.1.2                    # ASGI server
channels==4.0.0                  # Handles WebSocket connections
channels-redis==4.1.0            # Redis channel layer for Django Channels

# Database
psycopg2-binary==2.9.9           # PostgreSQL adapter
dj-database-url==2.1.0           # Database URL configuration

# Environment variables
python-dotenv==1.0.0             # Load environment variables from .env file

# Development tools (optional)
django-debug-toolbar==4.2.0       # Debug toolbar for development
